+++ reduct_cli/cli.py
@@ -9,9 +9,9 @@ from reduct_cli.config import write_config
 from reduct_cli.alias import alias
 from reduct_cli.bucket import bucket
 from reduct_cli.server import server
-from reduct_cli.mirror import mirror
 from reduct_cli.token import token
 from reduct_cli.export import export
+from reduct_cli.mirror import mirror
 
 
 @click.group()
@@ -62,6 +62,6 @@ def cli(
 cli.add_command(alias, "alias")
 cli.add_command(bucket, "bucket")
 cli.add_command(server, "server")
-cli.add_command(mirror, "mirror")
 cli.add_command(token, "token")
 cli.add_command(export, "export")
+cli.add_command(mirror, "mirror")