+++ reduct_cli/export.py
@@ -1,22 +1,34 @@
 """Export Command"""
-import asyncio
 from asyncio import new_event_loop as loop
-from pathlib import Path
 from typing import Optional
 
 import click
-from reduct import Client as ReductClient, Bucket, EntryInfo
-from rich.progress import Progress
+from reduct import Client as ReductClient
 
+from reduct_cli.export_impl.folder import export_to_folder
+from reduct_cli.export_impl.bucket import export_to_bucket
 from reduct_cli.utils.error import error_handle
 from reduct_cli.utils.helpers import (
     parse_path,
     get_alias,
-    read_records_with_progress,
-    filter_entries,
 )
 
 run = loop().run_until_complete
+
+start_option = click.option(
+    "--start",
+    help="Mirror records with timestamps newer than this time point in ISO format",
+)
+
+stop_option = click.option(
+    "--stop",
+    help="Mirror records  with timestamps older than this time point in ISO format",
+)
+entries_option = click.option(
+    "--entries",
+    help="Mirror only these entries, separated by comma",
+    default="",
+)
 
 
 @click.group()
@@ -24,58 +36,12 @@ def export():
     """Export data from a bucket somewhere else"""
 
 
-async def _export_entry(
-    path: Path, entry: EntryInfo, bucket: Bucket, progress: Progress, sem, **kwargs
-) -> None:
-    entry_path = Path(path / entry.name)
-    entry_path.mkdir(exist_ok=True)
-
-    async for record in read_records_with_progress(
-        entry, bucket, progress, sem, **kwargs
-    ):
-        with open(entry_path / f"{record.timestamp}.bin", "wb") as file:
-            async for chunk in record.read(1024):
-                file.write(chunk)
-
-
-async def _export_bucket(
-    client: ReductClient,
-    dest: str,
-    bucket_name: str,
-    parallel: int,
-    **kwargs,
-) -> None:
-    bucket: Bucket = await client.get_bucket(bucket_name)
-    folder_path = Path(dest) / bucket_name
-    folder_path.mkdir(parents=True, exist_ok=True)
-    sem = asyncio.Semaphore(parallel)
-
-    with Progress() as progress:
-        tasks = [
-            _export_entry(folder_path, entry, bucket, progress, sem, **kwargs)
-            for entry in filter_entries(
-                await bucket.get_entry_list(), kwargs["entries"]
-            )
-        ]
-        await asyncio.gather(*tasks)
-
-
 @export.command()
 @click.argument("src")
 @click.argument("dest")
-@click.option(
-    "--start",
-    help="Mirror records with timestamps newer than this time point in ISO format",
-)
-@click.option(
-    "--stop",
-    help="Mirror records  with timestamps older than this time point in ISO format",
-)
-@click.option(
-    "--entries",
-    help="Mirror only these entries, separated by comma",
-    default="",
-)
+@stop_option
+@start_option
+@entries_option
 @click.pass_context
 def folder(
     ctx, src: str, dest: str, start: Optional[str], stop: Optional[str], entries: str
@@ -83,20 +49,65 @@ def folder(
     """Export data from SRC bucket to DST folder
 
     SRC should be in the format of ALIAS/BUCKET_NAME
+    DST should be a path to a folder
     """
 
     with error_handle():
-        alias_name, bucket = parse_path(src)
+        alias_name, src_bucket = parse_path(src)
         alias = get_alias(ctx.obj["config_path"], alias_name)
 
         client = ReductClient(
             alias["url"], api_token=alias["token"], timeout=ctx.obj["timeout"]
         )
         run(
-            _export_bucket(
+            export_to_folder(
                 client,
                 dest,
-                bucket,
+                src_bucket,
+                parallel=ctx.obj["parallel"],
+                start=start,
+                stop=stop,
+                entries=entries.split(","),
+            )
+        )
+
+
+@export.command
+@click.argument("src")
+@click.argument("dest")
+@stop_option
+@start_option
+@entries_option
+@click.pass_context
+def bucket(
+    ctx, src: str, dest: str, start: Optional[str], stop: Optional[str], entries: str
+):  # pylint: disable=too-many-arguments
+    """Copy data from SRC to DEST bucket
+
+    SRC and DST should be in the format of ALIAS/BUCKET_NAME
+
+    If the destination bucket doesn't exist, it is created with
+    the settings of the source bucket."""
+
+    with error_handle():
+        alias_name, src_bucket = parse_path(src)
+        alias = get_alias(ctx.obj["config_path"], alias_name)
+        src_instance = ReductClient(
+            alias["url"], api_token=alias["token"], timeout=ctx.obj["timeout"]
+        )
+
+        alias_name, dest_bucket = parse_path(dest)
+        alias = get_alias(ctx.obj["config_path"], alias_name)
+        dest_instance = ReductClient(
+            alias["url"], api_token=alias["token"], timeout=ctx.obj["timeout"]
+        )
+
+        run(
+            export_to_bucket(
+                src_bucket,
+                dest_bucket,
+                src_instance,
+                dest_instance,
                 parallel=ctx.obj["parallel"],
                 start=start,
                 stop=stop,